+++ cmd/crypt/main.go
@@ -23,11 +23,13 @@ import (
 func main() {
 	if len(os.Args) < 2 {
 		fmt.Fprintln(os.Stderr, "usage: cryptd <command> [args]")
-		fmt.Fprintln(os.Stderr, "commands: solo, headless, autoplay, validate")
+		fmt.Fprintln(os.Stderr, "commands: serve, solo, headless, autoplay, validate")
 		os.Exit(1)
 	}
 
 	switch os.Args[1] {
+	case "serve":
+		runServe(os.Args[2:])
 	case "solo":
 		runSolo(os.Args[2:])
 	case "headless":